Fix "no scrolling to definition"
I run ctags on large code base (chromium). For some reason, if the file containing the definition was not already open, navigating to it would always scroll to the bottom of the file, and not to definition. Changing timeout to 2 from 1 fixes it, I do not know why.
My setup: ST3, Linux, Exuberant Ctags 5.9